==> 1. environment setup.R <==
# install under-development packaged used in the analysis
devtools::install("packages/InSituType2.0")
devtools::install("packages/smiDE_dev")
devtools::install_github("https://github.com/Nanostring-Biostats/insitucor")
